requirements.txt
openai>=1.0.0
python-docx>=0.8.11
reportlab>=3.6.1 